rework FoodItem
- rename to FoodItemRow
- extract the icons to their own component
